task_backend/requirements.txt
Django==4.2
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.1
djangorestframework==3.15.2
psycopg2==2.9.10
drf-yasg==1.21.8
django-cors-headers==4.7.0